95% done

+++ src/main/java/tn/esprit/spring/wecare/entities/BestAndWorstPost.java
@@ -2,6 +2,6 @@ package tn.esprit.spring.wecare.entities;
 
 public interface BestAndWorstPost {
 	public String getname();
-	public int getocc();
+	public double getocc();
 }
